+++ NEWS.md
@@ -1,5 +1,9 @@
 # parameters 0.8.7
 
+## New supported model classes
+
+* Preliminary support for `emm_list` objects.
+
 ## Changes to functions
 
 * `print()` for `model_parameters()` now names the coefficients column depending on the model type (i.e. `"Odds Ratios"` for logistic regression when `exponentiate = TRUE` etc.)